Get tests to pass
The incorrectly named class has been deleted, as this was causing
duplicate steps to be found and so the related tests were not running.
Once removed, there were two failures.

Bad Request was checking for an incorrect response code.

MadeUp was looking for a 200 response, but it actually returns a 404.
Spec has been amended to test this behaviour, but that is not to say it
is correct. Since correct behaviour cannot be determined from the docs
(since it always gives the 404 page), it will require a discussion.